move xdp setup to separate file
